Fixed box search and added comments and docs
I added quite a lot of documentation, so everything is now a bit easier to understand, hopefully. Users can now also search for objects with a box search. They can specify the width and the length of the box. I will have to check if the values I am using now are sort of correct